Remade the whole ECS
Remade the old, slow, and tedious to work ECS with a new one. There is still room for improvements in performance and interface, so changes are going to come in the near future. The system's part is also missing but with a better entity and component implementation that section will be easier to do. This ECS rework has been possible thanks to this article: austinmorlan.com/posts/entity_component_system