Add glm math library
Although I planned to create my own math library from scrath I decided to scrap the idea because of SIMD. This is quite heavy to implement because each compiler has a different aproach. Without this optimization the math library is so slow that it's best to discard it. So if in the future I want to learn more about this I will make my own implementation but for now I have decided to use glm.