Solve core dumped error on linux
Solved an error related to utilizing pointers when already freed that
affected linux. Th error happened because of storing all the events that
were emitted, which caused all of them to call the logger when being
deleted by the shared pointers, which was too late, as th elogger had
already been deleted. So, now the events are not stored and deleted
almost immediatly.